[Del] #27 - 파일 삭제
